Refactor the method for compiling XML
This one took me well over a day to complete. I had to create a bunch of
methods to break up the functionality.

Only one more to go, but I'm beginning to wonder if my refactoring is even
worth it, as the file is now over twice the starting size.